requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.12.0
evaluate>=0.4.0
PyYAML>=6.0.0

# Data processing
numpy>=1.21.0
pandas>=1.5.0
Pillow>=9.0.0
opencv-python>=4.7.0

# Machine learning
scikit-learn>=1.3.0
scipy>=1.10.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.15.0

# Experiment tracking
wandb>=0.15.0
tensorboard>=2.13.0

# Utilities
tqdm>=4.65.0
requests>=2.28.0
huggingface-hub>=0.16.0

# Development and testing
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
jupyter>=1.0.0

# Optional: For advanced image processing
albumentations>=1.3.0
imgaug>=0.4.0

# Optional: For model optimization
optuna>=3.2.0
